the angular admin forms are successfully connected to SwaggerOpenApi server, included its component template. Also angular material is installed and all the dependencies are updated.

+++ E-Commerce/Models/Product.cs
@@ -13,7 +13,7 @@ namespace E_Commerce.Models
         public string Description { get; set; } = null!;
 
         public decimal? Price { get; set; }
-        public byte? CurrencyId { get; set; }
+        public int? CurrencyId { get; set; }
         public decimal? Discount { get; set; }
         public short? QuantityAvailable { get; set; }
